# internal/interfaces/http/handler.go
package http

import (
	"encoding/json"
	"net/http"
	"strconv"

	"github.com/gin-gonic/gin"
	"github.com/go-chi/chi/v5"
	httpSwagger "github.com/swaggo/http-swagger/v2"
	"go.uber.org/zap"

	_ "github.com/himmel/order-service/docs" // docs generated by Swag CLI
	"github.com/himmel/order-service/internal/application/command"
	"github.com/himmel/order-service/internal/application/dto"
	"github.com/himmel/order-service/internal/application/query"
)

// Handler holds all HTTP handlers and their dependencies.
type Handler struct {
	createOrder  *command.CreateOrderHandler
	addItem      *command.AddItemHandler
	confirmOrder *command.ConfirmOrderHandler
	getOrder     *query.GetOrderHandler
	listOrders   *query.ListOrdersHandler
	validator    *RequestValidator
	logger       *zap.SugaredLogger
}

func NewHandler(
	createOrder *command.CreateOrderHandler,
	addItem *command.AddItemHandler,
	confirmOrder *command.ConfirmOrderHandler,
	getOrder *query.GetOrderHandler,
	listOrders *query.ListOrdersHandler,
	logger *zap.SugaredLogger,
) *Handler {
	return &Handler{
		createOrder:  createOrder,
		addItem:      addItem,
		confirmOrder: confirmOrder,
		getOrder:     getOrder,
		listOrders:   listOrders,
		validator:    NewRequestValidator(),
		logger:       logger,
	}
}

// RegisterRoutes sets up chi routes.
func (h *Handler) RegisterRoutes(r chi.Router) {
	// Swagger UI
	r.Get("/swagger/*", httpSwagger.Handler(
		httpSwagger.URL("/swagger/doc.json"), //The url pointing to API definition
	))

	r.Route("/orders", func(r chi.Router) {
		r.Post("/", h.HandleCreateOrder)
		r.Get("/", h.HandleListOrders)
		r.Route("/{orderID}", func(r chi.Router) {
			r.Get("/", h.HandleGetOrder)
			r.Post("/items", h.HandleAddItem)
			r.Post("/confirm", h.HandleConfirmOrder)
		})
	})
}

// @Summary Create a new order
// @Description Create an order for a given customer
// @Tags orders
// @Accept json
// @Produce json
// @Param request body dto.CreateOrderRequest true "Create Order Request"
// @Success 201 {object} dto.CreateOrderResult
// @Failure 400 {object} dto.ErrorResponse
// @Failure 500 {object} dto.ErrorResponse
// @Router /orders [post]
func (h *Handler) HandleCreateOrder(w http.ResponseWriter, r *http.Request) {
	var req dto.CreateOrderRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		writeError(w, r, http.StatusBadRequest, "INVALID_REQUEST", "Invalid JSON body", nil)
		return
	}

	if errs := h.validator.Validate(req); errs != nil {
		writeError(w, r, http.StatusBadRequest, "VALIDATION_ERROR", "Request validation failed", errs)
		return
	}

	orderID, err := h.createOrder.Handle(r.Context(), command.CreateOrder{
		CustomerID: req.CustomerID,
	})
	if err != nil {
		mapAndWriteError(w, r, err)
		return
	}

	writeJSON(w, http.StatusCreated, dto.CreateOrderResult{OrderID: orderID})
}

// @Summary Add an item to an order
// @Description Add a new product to an existing DRAFT order
// @Tags orders
// @Accept json
// @Produce json
// @Param orderID path string true "Order ID"
// @Param request body dto.AddItemRequest true "Add Item Request"
// @Success 200 {object} map[string]string
// @Failure 400 {object} dto.ErrorResponse
// @Failure 404 {object} dto.ErrorResponse
// @Failure 409 {object} dto.ErrorResponse
// @Failure 500 {object} dto.ErrorResponse
// @Router /orders/{orderID}/items [post]
func (h *Handler) HandleAddItem(w http.ResponseWriter, r *http.Request) {
	orderID := chi.URLParam(r, "orderID")

	var req dto.AddItemRequest
	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
		writeError(w, r, http.StatusBadRequest, "INVALID_REQUEST", "Invalid JSON body", nil)
		return
	}

	if errs := h.validator.Validate(req); errs != nil {
		writeError(w, r, http.StatusBadRequest, "VALIDATION_ERROR", "Request validation failed", errs)
		return
	}

	err := h.addItem.Handle(r.Context(), command.AddItemToOrder{
		OrderID:     orderID,
		ProductID:   req.ProductID,
		ProductName: req.ProductName,
		Quantity:    req.Quantity,
		UnitPrice:   req.UnitPrice,
	})
	if err != nil {
		mapAndWriteError(w, r, err)
		return
	}

	writeJSON(w, http.StatusOK, map[string]string{"message": "item added"})
}

// @Summary Confirm an order
// @Description Transition order status from DRAFT to CONFIRMED
// @Tags orders
// @Accept json
// @Produce json
// @Param orderID path string true "Order ID"
// @Success 200 {object} map[string]string
// @Failure 400 {object} dto.ErrorResponse
// @Failure 404 {object} dto.ErrorResponse
// @Failure 409 {object} dto.ErrorResponse
// @Failure 500 {object} dto.ErrorResponse
// @Router /orders/{orderID}/confirm [post]
func (h *Handler) HandleConfirmOrder(w http.ResponseWriter, r *http.Request) {
	orderID := chi.URLParam(r, "orderID")

	err := h.confirmOrder.Handle(r.Context(), command.ConfirmOrder{
		OrderID: orderID,
	})
	if err != nil {
		mapAndWriteError(w, r, err)
		return
	}

	writeJSON(w, http.StatusOK, map[string]string{"message": "order confirmed"})
}

// @Summary Get order details
// @Description Fetch an order by its ID from the read model
// @Tags orders
// @Accept json
// @Produce json
// @Param orderID path string true "Order ID"
// @Success 200 {object} dto.OrderResponse
// @Failure 404 {object} dto.ErrorResponse
// @Failure 500 {object} dto.ErrorResponse
// @Router /orders/{orderID} [get]
func (h *Handler) HandleGetOrder(w http.ResponseWriter, r *http.Request) {
	orderID := chi.URLParam(r, "orderID")

	result, err := h.getOrder.Handle(r.Context(), query.GetOrderByID{
		OrderID: orderID,
	})
	if err != nil {
		mapAndWriteError(w, r, err)
		return
	}

	writeJSON(w, http.StatusOK, result)
}

// @Summary List orders
// @Description Get a paginated list of orders, optionally filtered
// @Tags orders
// @Accept json
// @Produce json
// @Param status query string false "Filter by order status"
// @Param customer_id query string false "Filter by customer ID"
// @Param page query int false "Page number" default(1)
// @Param size query int false "Page size" default(20)
// @Param sort query string false "Sort field (created_at, total_amount)"
// @Param dir query string false "Sort direction (asc, desc)"
// @Success 200 {object} dto.PaginatedOrdersResponse
// @Failure 500 {object} dto.ErrorResponse
// @Router /orders [get]
func (h *Handler) HandleListOrders(w http.ResponseWriter, r *http.Request) {
	q := query.ListOrders{
		SortBy:   r.URL.Query().Get("sort"),
		SortDir:  r.URL.Query().Get("dir"),
		Page:     1,
		PageSize: 20,
	}

	if v := r.URL.Query().Get("status"); v != "" {
		q.Status = &v
	}
	if v := r.URL.Query().Get("customer_id"); v != "" {
		q.CustomerID = &v
	}
	if v := r.URL.Query().Get("page"); v != "" {
		if p, err := strconv.Atoi(v); err == nil && p > 0 {
			q.Page = p
		}
	}
	if v := r.URL.Query().Get("size"); v != "" {
		if s, err := strconv.Atoi(v); err == nil && s > 0 {
			q.PageSize = s
		}
	}

	result, err := h.listOrders.Handle(r.Context(), q)
	if err != nil {
		mapAndWriteError(w, r, err)
		return
	}

	writeJSON(w, http.StatusOK, result)
}

func writeJSON(w http.ResponseWriter, status int, data interface{}) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	json.NewEncoder(w).Encode(data)
}

// ============ Gin Handler Methods ============

// GinHandleCreateOrder handles POST /orders with Gin.
// @Summary Create a new order
// @Description Create an order for a given customer
// @Tags orders
// @Accept json
// @Produce json
// @Param request body dto.CreateOrderRequest true "Create Order Request"
// @Success 201 {object} dto.CreateOrderResult
// @Failure 400 {object} dto.ErrorResponse
// @Failure 500 {object} dto.ErrorResponse
// @Router /orders [post]
func (h *Handler) GinHandleCreateOrder(c *gin.Context) {
	var req dto.CreateOrderRequest
	if err := c.BindJSON(&req); err != nil {
		h.logger.Debugw("invalid request body", "error", err)
		c.JSON(http.StatusBadRequest, dto.ErrorResponse{
			Code:    "INVALID_REQUEST",
			Message: "Invalid JSON body",
		})
		return
	}

	if errs := h.validator.Validate(req); errs != nil {
		h.logger.Debugw("validation failed", "validation errors", errs)
		c.JSON(http.StatusBadRequest, dto.ErrorResponse{
			Code:    "VALIDATION_ERROR",
			Message: "Request validation failed",
			Details: errs,
		})
		return
	}

	h.logger.Infow("creating order", "customer_id", req.CustomerID)
	orderID, err := h.createOrder.Handle(c.Request.Context(), command.CreateOrder{
		CustomerID: req.CustomerID,
	})
	if err != nil {
		h.mapAndWriteGinError(c, err)
		return
	}

	h.logger.Infow("order created successfully", "order_id", orderID, "customer_id", req.CustomerID)
	c.JSON(http.StatusCreated, dto.CreateOrderResult{OrderID: orderID})
}

// GinHandleAddItem handles POST /orders/{orderID}/items with Gin.
// @Summary Add an item to an order
// @Description Add a new product to an existing DRAFT order
// @Tags orders
// @Accept json
// @Produce json
// @Param orderID path string true "Order ID"
// @Param request body dto.AddItemRequest true "Add Item Request"
// @Success 200 {object} map[string]string
// @Failure 400 {object} dto.ErrorResponse
// @Failure 404 {object} dto.ErrorResponse
// @Failure 409 {object} dto.ErrorResponse
// @Failure 500 {object} dto.ErrorResponse
// @Router /orders/{orderID}/items [post]
func (h *Handler) GinHandleAddItem(c *gin.Context) {
	orderID := c.Param("orderID")

	var req dto.AddItemRequest
	if err := c.BindJSON(&req); err != nil {
		h.logger.Debugw("invalid request body", "error", err, "order_id", orderID)
		c.JSON(http.StatusBadRequest, dto.ErrorResponse{
			Code:    "INVALID_REQUEST",
			Message: "Invalid JSON body",
		})
		return
	}

	if errs := h.validator.Validate(req); errs != nil {
		h.logger.Debugw("validation failed", "order_id", orderID, "validation errors", errs)
		c.JSON(http.StatusBadRequest, dto.ErrorResponse{
			Code:    "VALIDATION_ERROR",
			Message: "Request validation failed",
			Details: errs,
		})
		return
	}

	h.logger.Infow("adding item to order",
		"order_id", orderID,
		"product_id", req.ProductID,
		"product_name", req.ProductName,
		"quantity", req.Quantity,
	)
	err := h.addItem.Handle(c.Request.Context(), command.AddItemToOrder{
		OrderID:     orderID,
		ProductID:   req.ProductID,
		ProductName: req.ProductName,
		Quantity:    req.Quantity,
		UnitPrice:   req.UnitPrice,
	})
	if err != nil {
		h.mapAndWriteGinError(c, err)
		return
	}

	h.logger.Infow("item added successfully", "order_id", orderID, "product_id", req.ProductID)
	c.JSON(http.StatusOK, map[string]string{"message": "item added"})
}

// GinHandleConfirmOrder handles POST /orders/{orderID}/confirm with Gin.
// @Summary Confirm an order
// @Description Transition order status from DRAFT to CONFIRMED
// @Tags orders
// @Accept json
// @Produce json
// @Param orderID path string true "Order ID"
// @Success 200 {object} map[string]string
// @Failure 400 {object} dto.ErrorResponse
// @Failure 404 {object} dto.ErrorResponse
// @Failure 409 {object} dto.ErrorResponse
// @Failure 500 {object} dto.ErrorResponse
// @Router /orders/{orderID}/confirm [post]
func (h *Handler) GinHandleConfirmOrder(c *gin.Context) {
	orderID := c.Param("orderID")

	h.logger.Infow("confirming order", "order_id", orderID)
	err := h.confirmOrder.Handle(c.Request.Context(), command.ConfirmOrder{
		OrderID: orderID,
	})
	if err != nil {
		h.mapAndWriteGinError(c, err)
		return
	}

	h.logger.Infow("order confirmed successfully", "order_id", orderID)
	c.JSON(http.StatusOK, map[string]string{"message": "order confirmed"})
}

// GinHandleGetOrder handles GET /orders/{orderID} with Gin.
// @Summary Get order details
// @Description Fetch an order by its ID from the read model
// @Tags orders
// @Accept json
// @Produce json
// @Param orderID path string true "Order ID"
// @Success 200 {object} dto.OrderResponse
// @Failure 404 {object} dto.ErrorResponse
// @Failure 500 {object} dto.ErrorResponse
// @Router /orders/{orderID} [get]
func (h *Handler) GinHandleGetOrder(c *gin.Context) {
	orderID := c.Param("orderID")

	h.logger.Debugw("fetching order", "order_id", orderID)
	result, err := h.getOrder.Handle(c.Request.Context(), query.GetOrderByID{
		OrderID: orderID,
	})
	if err != nil {
		h.mapAndWriteGinError(c, err)
		return
	}

	h.logger.Debugw("order fetched successfully", "order_id", orderID)
	c.JSON(http.StatusOK, result)
}

// GinHandleListOrders handles GET /orders with Gin.
// @Summary List orders
// @Description Get a paginated list of orders, optionally filtered
// @Tags orders
// @Accept json
// @Produce json
// @Param status query string false "Filter by order status"
// @Param customer_id query string false "Filter by customer ID"
// @Param page query int false "Page number" default(1)
// @Param size query int false "Page size" default(20)
// @Param sort query string false "Sort field (created_at, total_amount)"
// @Param dir query string false "Sort direction (asc, desc)"
// @Success 200 {object} dto.PaginatedOrdersResponse
// @Failure 500 {object} dto.ErrorResponse
// @Router /orders [get]
func (h *Handler) GinHandleListOrders(c *gin.Context) {
	q := query.ListOrders{
		SortBy:   c.DefaultQuery("sort", "created_at"),
		SortDir:  c.DefaultQuery("dir", "desc"),
		Page:     1,
		PageSize: 20,
	}

	if v := c.Query("status"); v != "" {
		q.Status = &v
	}
	if v := c.Query("customer_id"); v != "" {
		q.CustomerID = &v
	}
	if v := c.Query("page"); v != "" {
		if p, err := strconv.Atoi(v); err == nil && p > 0 {
			q.Page = p
		}
	}
	if v := c.Query("size"); v != "" {
		if s, err := strconv.Atoi(v); err == nil && s > 0 {
			q.PageSize = s
		}
	}

	h.logger.Debugw("listing orders",
		"page", q.Page,
		"page_size", q.PageSize,
		"sort_by", q.SortBy,
		"sort_dir", q.SortDir,
	)
	result, err := h.listOrders.Handle(c.Request.Context(), q)
	if err != nil {
		h.mapAndWriteGinError(c, err)
		return
	}

	h.logger.Debugw("orders listed successfully",
		"page", q.Page,
		"total_items", len(result.Orders),
	)
	c.JSON(http.StatusOK, result)
}

// RegisterGinRoutes registers all Gin routes.
func (h *Handler) RegisterGinRoutes(r *gin.Engine) {
	// Swagger UI - wrap the http.HandlerFunc for Gin
	r.GET("/swagger/*any", gin.WrapF(httpSwagger.WrapHandler))

	// Health endpoints (not in handler group)
	orders := r.Group("/orders")
	{
		orders.POST("", h.GinHandleCreateOrder)
		orders.GET("", h.GinHandleListOrders)

		orders.GET("/:orderID", h.GinHandleGetOrder)
		orders.POST("/:orderID/items", h.GinHandleAddItem)
		orders.POST("/:orderID/confirm", h.GinHandleConfirmOrder)
	}
}

// mapAndWriteGinError maps domain errors to HTTP responses for Gin.
func (h *Handler) mapAndWriteGinError(c *gin.Context, err error) {
	h.logger.Errorw("handler error",
		"error", err.Error(),
		"type", "domain_error",
	)
	mapAndWriteError(c.Writer, c.Request, err)
}